requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
psycopg2-binary==2.9.9
